requirements.txt
python==3.9.13
django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
django-rest-auth==0.9.5
django-allauth==0.51.0